+++ src/pages/catalog/components/catalogFilter/components/CustomCheckbox.tsx
@@ -1,5 +1,5 @@
 import { Path, useFormContext, UseFormRegister } from 'react-hook-form';
-import { Filters } from '../catalog-filter.tsx';
+import { Filters } from '../CatalogFilter.tsx';
 
 export function CustomCheckbox({
   register,
