+++ src/pages/catalog/components/catalogFilter/components/GetFilterResetSvg.tsx
